requirements.txt
# Core dependencies for MCD PoC
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
